+++ public/index.html
@@ -2,6 +2,7 @@
 <html lang="en">
 
 <head>
+  <script type="text/javascript" src="//dapi.kakao.com/v2/maps/sdk.js?appkey=1f4f02f2debec9f216d1f79e27073fef"></script>
   <meta charset="utf-8" />
   <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
   <meta name="viewport" content="width=device-width, initial-scale=1" />
@@ -10,11 +11,10 @@
   <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
   <meta name="viewport" content="width=1024, initial-scale=1">
 
-  <!--
-      manifest.json provides metadata used when your web app is installed on a
-      user's mobile device or desktop. See https://developers.google.com/web/fundamentals/web-app-manifest/
-    -->
   <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
+
+
+
   <!--
       Notice the use of %PUBLIC_URL% in the tags above.
       It will be replaced with the URL of the `public` folder during the build.
